+++ Lib9c.Tools.Tests/Data/TableCSV/Cost/EnhancementCostSheet.csv
@@ -1,0 +1,251 @@
+id,item_sub_type,grade,level,cost
+1,Weapon,1,1,0
+2,Weapon,1,2,0
+3,Weapon,1,3,0
+4,Weapon,1,4,1000
+5,Weapon,1,5,0
+6,Weapon,1,6,0
+7,Weapon,1,7,5000
+8,Weapon,1,8,0
+9,Weapon,1,9,0
+10,Weapon,1,10,25000
+11,Weapon,2,1,0
+12,Weapon,2,2,0
+13,Weapon,2,3,0
+14,Weapon,2,4,2000
+15,Weapon,2,5,0
+16,Weapon,2,6,0
+17,Weapon,2,7,10000
+18,Weapon,2,8,0
+19,Weapon,2,9,0
+20,Weapon,2,10,50000
+21,Weapon,3,1,0
+22,Weapon,3,2,0
+23,Weapon,3,3,0
+24,Weapon,3,4,3000
+25,Weapon,3,5,0
+26,Weapon,3,6,0
+27,Weapon,3,7,15000
+28,Weapon,3,8,0
+29,Weapon,3,9,0
+30,Weapon,3,10,75000
+31,Weapon,4,1,0
+32,Weapon,4,2,0
+33,Weapon,4,3,0
+34,Weapon,4,4,6000
+35,Weapon,4,5,0
+36,Weapon,4,6,0
+37,Weapon,4,7,30000
+38,Weapon,4,8,0
+39,Weapon,4,9,0
+40,Weapon,4,10,150000
+41,Weapon,5,1,0
+42,Weapon,5,2,0
+43,Weapon,5,3,0
+44,Weapon,5,4,9000
+45,Weapon,5,5,0
+46,Weapon,5,6,0
+47,Weapon,5,7,45000
+48,Weapon,5,8,0
+49,Weapon,5,9,0
+50,Armor,5,10,225000
+51,Armor,1,1,0
+52,Armor,1,2,0
+53,Armor,1,3,0
+54,Armor,1,4,1000
+55,Armor,1,5,0
+56,Armor,1,6,0
+57,Armor,1,7,5000
+58,Armor,1,8,0
+59,Armor,1,9,0
+60,Armor,1,10,25000
+61,Armor,2,1,0
+62,Armor,2,2,0
+63,Armor,2,3,0
+64,Armor,2,4,2000
+65,Armor,2,5,0
+66,Armor,2,6,0
+67,Armor,2,7,10000
+68,Armor,2,8,0
+69,Armor,2,9,0
+70,Armor,2,10,50000
+71,Armor,3,1,0
+72,Armor,3,2,0
+73,Armor,3,3,0
+74,Armor,3,4,3000
+75,Armor,3,5,0
+76,Armor,3,6,0
+77,Armor,3,7,15000
+78,Armor,3,8,0
+79,Armor,3,9,0
+80,Armor,3,10,75000
+81,Armor,4,1,0
+82,Armor,4,2,0
+83,Armor,4,3,0
+84,Armor,4,4,6000
+85,Armor,4,5,0
+86,Armor,4,6,0
+87,Armor,4,7,30000
+88,Armor,4,8,0
+89,Armor,4,9,0
+90,Armor,4,10,150000
+91,Armor,5,1,0
+92,Armor,5,2,0
+93,Armor,5,3,0
+94,Armor,5,4,9000
+95,Armor,5,5,0
+96,Armor,5,6,0
+97,Armor,5,7,45000
+98,Armor,5,8,0
+99,Armor,5,9,0
+100,Armor,5,10,225000
+101,Belt,1,1,0
+102,Belt,1,2,0
+103,Belt,1,3,0
+104,Belt,1,4,1000
+105,Belt,1,5,0
+106,Belt,1,6,0
+107,Belt,1,7,5000
+108,Belt,1,8,0
+109,Belt,1,9,0
+110,Belt,1,10,25000
+111,Belt,2,1,0
+112,Belt,2,2,0
+113,Belt,2,3,0
+114,Belt,2,4,2000
+115,Belt,2,5,0
+116,Belt,2,6,0
+117,Belt,2,7,10000
+118,Belt,2,8,0
+119,Belt,2,9,0
+120,Belt,2,10,50000
+121,Belt,3,1,0
+122,Belt,3,2,0
+123,Belt,3,3,0
+124,Belt,3,4,3000
+125,Belt,3,5,0
+126,Belt,3,6,0
+127,Belt,3,7,15000
+128,Belt,3,8,0
+129,Belt,3,9,0
+130,Belt,3,10,75000
+131,Belt,4,1,0
+132,Belt,4,2,0
+133,Belt,4,3,0
+134,Belt,4,4,6000
+135,Belt,4,5,0
+136,Belt,4,6,0
+137,Belt,4,7,30000
+138,Belt,4,8,0
+139,Belt,4,9,0
+140,Belt,4,10,150000
+141,Belt,5,1,0
+142,Belt,5,2,0
+143,Belt,5,3,0
+144,Belt,5,4,9000
+145,Belt,5,5,0
+146,Belt,5,6,0
+147,Belt,5,7,45000
+148,Belt,5,8,0
+149,Belt,5,9,0
+150,Belt,5,10,225000
+151,Necklace,1,1,0
+152,Necklace,1,2,0
+153,Necklace,1,3,0
+154,Necklace,1,4,1000
+155,Necklace,1,5,0
+156,Necklace,1,6,0
+157,Necklace,1,7,5000
+158,Necklace,1,8,0
+159,Necklace,1,9,0
+160,Necklace,1,10,25000
+161,Necklace,2,1,0
+162,Necklace,2,2,0
+163,Necklace,2,3,0
+164,Necklace,2,4,2000
+165,Necklace,2,5,0
+166,Necklace,2,6,0
+167,Necklace,2,7,10000
+168,Necklace,2,8,0
+169,Necklace,2,9,0
+170,Necklace,2,10,50000
+171,Necklace,3,1,0
+172,Necklace,3,2,0
+173,Necklace,3,3,0
+174,Necklace,3,4,3000
+175,Necklace,3,5,0
+176,Necklace,3,6,0
+177,Necklace,3,7,15000
+178,Necklace,3,8,0
+179,Necklace,3,9,0
+180,Necklace,3,10,75000
+181,Necklace,4,1,0
+182,Necklace,4,2,0
+183,Necklace,4,3,0
+184,Necklace,4,4,6000
+185,Necklace,4,5,0
+186,Necklace,4,6,0
+187,Necklace,4,7,30000
+188,Necklace,4,8,0
+189,Necklace,4,9,0
+190,Necklace,4,10,150000
+191,Necklace,5,1,0
+192,Necklace,5,2,0
+193,Necklace,5,3,0
+194,Necklace,5,4,9000
+195,Necklace,5,5,0
+196,Necklace,5,6,0
+197,Necklace,5,7,45000
+198,Necklace,5,8,0
+199,Necklace,5,9,0
+200,Necklace,5,10,225000
+201,Ring,1,1,0
+202,Ring,1,2,0
+203,Ring,1,3,0
+204,Ring,1,4,1000
+205,Ring,1,5,0
+206,Ring,1,6,0
+207,Ring,1,7,5000
+208,Ring,1,8,0
+209,Ring,1,9,0
+210,Ring,1,10,25000
+211,Ring,2,1,0
+212,Ring,2,2,0
+213,Ring,2,3,0
+214,Ring,2,4,2000
+215,Ring,2,5,0
+216,Ring,2,6,0
+217,Ring,2,7,10000
+218,Ring,2,8,0
+219,Ring,2,9,0
+220,Ring,2,10,50000
+221,Ring,3,1,0
+222,Ring,3,2,0
+223,Ring,3,3,0
+224,Ring,3,4,3000
+225,Ring,3,5,0
+226,Ring,3,6,0
+227,Ring,3,7,15000
+228,Ring,3,8,0
+229,Ring,3,9,0
+230,Ring,3,10,75000
+231,Ring,4,1,0
+232,Ring,4,2,0
+233,Ring,4,3,0
+234,Ring,4,4,6000
+235,Ring,4,5,0
+236,Ring,4,6,0
+237,Ring,4,7,30000
+238,Ring,4,8,0
+239,Ring,4,9,0
+240,Ring,4,10,150000
+241,Ring,5,1,0
+242,Ring,5,2,0
+243,Ring,5,3,0
+244,Ring,5,4,9000
+245,Ring,5,5,0
+246,Ring,5,6,0
+247,Ring,5,7,45000
+248,Ring,5,8,0
+249,Ring,5,9,0
+250,Ring,5,10,225000
