+++ Lib9c.Tools.Tests/Data/TableCSV/GameConfigSheet.csv
@@ -1,0 +1,3 @@
+key,value
+hourglass_per_block,3
+action_point_max,120
